% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatnet.R
\name{neat_geometry_buffer}
\alias{neat_geometry_buffer}
\title{Buffer and Union Lines}
\usage{
neat_geometry_buffer(geometry, dist)
}
\arguments{
\item{geometry}{A geos_geometry vector of lines.}

\item{dist}{Buffer distance.}
}
\value{
A geos_geometry vector (polygons).
}
\description{
Buffers a set of lines and unions them into a single polygon.
}
