% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatnet.R
\name{neat_skeletonize}
\alias{neat_skeletonize}
\title{Skeletonize Polygon}
\usage{
neat_skeletonize(
  geometry,
  dist,
  max_segment_length,
  final_min_length,
  simplify = FALSE
)
}
\arguments{
\item{geometry}{A geos_geometry vector (polygons).}

\item{dist}{The buffer distance used for the input geometry (needed for final pruning).}

\item{max_segment_length}{Distance for densifying the boundary to ensure good Voronoi generation.}

\item{final_min_length}{Minimum length for final segment filtering.}
}
\value{
A geos_geometry vector (lines).
}
\description{
Creates a skeleton of a polygon using Voronoi diagrams.
}
