% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatnet.R
\name{neatnet}
\alias{neatnet}
\title{Neat Network Simplification}
\usage{
neatnet(x, dist = 10, max_segment_factor = 2, final_min_factor = 3)
}
\arguments{
\item{x}{An sf object representing the network.}

\item{dist}{Buffer distance (half width of road).}

\item{max_segment_factor}{Multiplier for dist to determine max_segment_length (default 2).}

\item{final_min_factor}{Multiplier for dist to determine final_min_length (default 3).}
}
\value{
An sf object of the simplified network.
}
\description{
Simplifies a road network using a buffer-based skeletonization approach.
}
